We can now edit and modify races and classes.
- Removed dumps from files.

+++ app/Flare/Models/GameRace.php
@@ -36,11 +36,11 @@ class GameRace extends Model
      * @var array
      */
     protected $casts = [
-        'str_mod'      => 'float',
-        'dur_mod'      => 'float',
-        'dex_mod'      => 'float',
-        'chr_mod'      => 'float',
-        'int_mod'      => 'float',
+        'str_mod'      => 'integer',
+        'dur_mod'      => 'integer',
+        'dex_mod'      => 'integer',
+        'chr_mod'      => 'integer',
+        'int_mod'      => 'integer',
         'accuracy_mod' => 'float',
         'dodge_mod'    => 'float',
         'deffense_mod' => 'float',
